Remove out of date libpostal settings
We have some settings hanging around our docker-compose.yml files that
are hanging around from when we did things quite differently.

Previously, we built the libpostal_baseimage image directly as part of
our Docker setup projects. This required setting it as a "dependency" of
other images. That's no longer required, and having that `dependency`
setting is causing an extra container to be launched, after which it
does nothing and shuts down.

Additionaly, we used to store libpostal training data in a separate
filesystem. It's now baked into the `libpostal_baseimage` image, so all
those settings are no longer required.