fix(ES image): Allow running Elasticsearch 5 container as any user
This sets some more permissive permissions on various directories,
files, and executables used in the Elasticsearch 5 docker image.

It allows the user the container processes run under to be changed from
the default of 1000, which helps keep things consistent with the rest of
the Pelias containers.

Fixes https://github.com/pelias/docker/issues/46