Add Elasticsearch 7.4.0 Docker image
This is the first step in supporting Elasticsearch 7.

At this time, Pelias does not work out of the box on ES7, but with a
Docker image ready to go, we can begin testing changes for
compatibility.

This Dockerfile and config is identical to the ES6 Docker image, except
for changing the version, and making one update to the
`elasticsearch.yml`:

In ES7, the bulk thread pool is removed, and both bulk and non-bulk
operations go through a single
[write](https://www.elastic.co/guide/en/elasticsearch/reference/current/modules-threadpool.html#modules-threadpool)
thread pool.

For Pelias we have found increasing the queue size of this thread pool
is useful to ensure imports can succeed without errors, so the
configuration file has been updated accordingly.

Connects https://github.com/pelias/pelias/issues/831